Fix json parsing of the `details` field in API error responses with a new `ErrorResponseDetail` type.
The `details` field in the json response from API errors is an array of json objects (not a single object).
Also, the value for the "link" object is not a string but rather another nested array of json objects.
The parsing of the `details` json field is now correctly unmarshalled into a new `ErrorResponseDetail` type
instead of the previously broken `map[string]string` type.