Added url and image_url for a item (#281)
When using CreateOrder we have PurchaseUnitRequest inside it
a list of Items, according to paypal docs
https://developer.paypal.com/docs/api/orders/v2/#orders_create

this item can have a url, and a image_url

url:
The URL to the item being purchased. Visible to buyer and used in buyer experiences.

image_url:
The URL of the item's image.